scrap-strava-data/requirements.txt
python-dateutil
python-dotenv
requests 
google-cloud-storage